feat: Add PyO3 bindings and regenerate Cargo.lock
- Add `pyo3` and `numpy` optional dependencies to `crates/geo-polygonize-core/Cargo.toml`.
- Introduce `python` feature in `crates/geo-polygonize-core`.
- Implement PyO3 `polygonize` function in `crates/geo-polygonize-core/src/python.rs`.
- Configure `pyproject.toml` for `maturin` build.
- Refactor python package structure to support both CFFI and PyO3:
  - Add `python/geo_polygonize/types.py` for `SimplePolygon`.
  - Add `python/geo_polygonize/cffi_wrapper.py` for CFFI fallback.
  - Update `python/geo_polygonize/__init__.py` to handle imports.
- Regenerate `Cargo.lock` via `cargo update`.
