+++ crates/geo-polygonize-core/Cargo.toml
@@ -17,6 +17,8 @@ log = "0.4"
 thiserror = "1.0"
 float_next_after = "1.0"
 smallvec = "1.11"
+pyo3 = { version = "0.21", features = ["extension-module", "abi3-py37"], optional = true }
+numpy = { version = "0.21", optional = true }
 
 [dev-dependencies]
 approx = "0.5"
@@ -30,6 +32,7 @@ serde_json = "1.0"
 [features]
 default = ["parallel"]
 parallel = ["dep:rayon"]
+python = ["dep:pyo3", "dep:numpy"]
 
 [[bench]]
 name = "polygonize_bench"
