+++ pyproject.toml
@@ -1,0 +1,19 @@
+[build-system]
+requires = ["maturin>=1.4,<2.0"]
+build-backend = "maturin"
+
+[project]
+name = "geo_polygonize"
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Rust",
+    "Programming Language :: Python :: Implementation :: CPython",
+    "Programming Language :: Python :: Implementation :: PyPy",
+]
+dependencies = ["numpy", "cffi"]
+
+[tool.maturin]
+python-source = "python"
+manifest-path = "crates/geo-polygonize-core/Cargo.toml"
+module-name = "geo_polygonize.geo_polygonize_core"
+features = ["python"]
